+++ DataModule/src/main/java/com/currency/exchange/datamodule/data/model/entities/RateDTO.kt
@@ -4,8 +4,8 @@ import androidx.room.Entity
 
 @Entity(tableName = "rates_dto", primaryKeys = ["base", "date"])
 data class RateDTO(
-    var amount: Double,
+    var amount: Float,
     val base: String,
     var date: String,
-    val rates: Map<String, Double>,
+    val rates: Map<String, Float>,
 )
